php7-ext: work around isnan absence on gcc5
In an environment with the combination: C++ >= 2011, glibc >= 2.23, g++ < 6,
"isnan" is not a recognized identifier. The reasons are multifold.

First, glibc-2.23 started to hide the isnan definition for C++2011:

	--- glibc-2.19
	+++ glibc-2.23
	-#if defined __USE_MISC || defined __USE_XOPEN
	-/* Return nonzero if VALUE is not a number.  */
	-__MATHDECL_1 (int,isnan,, (_Mdouble_ __value)) __attribute__ ((__const__));
	[...]
	+#if defined __USE_MISC || (defined __USE_XOPEN && !defined __USE_XOPEN2K)
	+# if (!defined __cplusplus \
	+      || __cplusplus < 201103L /* isnan conflicts with C++11.  */ \
	+      || __MATH_DECLARING_DOUBLE == 0) /* isnanf or isnanl don't.  */
	+/* Return nonzero if VALUE is not a number.  */
	+__MATHDECL_1 (int,isnan,, (_Mdouble_ __value)) __attribute__ ((__const__));
	+# endif
	+#endif

Second, g++ has added an import for it again in the file /usr/include/c++/6/math.h,
but which only exists in g++6 onwards and not in previous versions.

	using std::isnan;

Since php may be relying on more of the names available in C++98, go for an
"using namespace std".