libicalmapi: improve fallback scenario to server_timezone in ical.cfg
Instead of taking the first timezone component of its kind from the
builtin timezones, loop over them and pick the one for the correct
period.

As these builtin timezones apparently to not have to contain
rrules/rdates (anymore?), continue without their presence (just use
dtstart).

Improve the check for 'Asia/Kolkata'-like timezones, so as to not
interfere with the above.

References: KC-11