libserver: let PRI keys not be NULL
All parts of PRIMARY keys need to be NOT NULL, otherwise a UNIQUE key
would be required.

Since we do not want (nor require) a UNIQUE key (specifically because
of the performance downsides), we therefore just update the table
definition to disallow NULL values, yet setting "0" as default value.

This commit might not be the last one regarding MySQL 5.7
compatibility.

References: KC-2