server: replace "REPLACE INTO" quries by "INSERT ON DUPLICATE KEY"
This commit replaces some core "REPLACE" statements by "INSERT ... ON
DUPLICATE KEY" statements — the benefit is the direct update based on
the matching key instead of (in case of REPLACE INTO) first deleting
the matching row and re-inserting it.

The earlier DELETE row operation issues an extra index update whereas
the INSERT ON DUPLICATE KEY does in fact an in-place update operation.

References: KC-58