server: avoid accessing uninitialized variable in error path
Cherry-pick commit kopanocore-8.0.0-91-gdcbd1ef .

"sObject" starts out uninitialized, and is either initialized by
GetCacheItem (if successful), or by "sObject = &sNewObject" — but only
if _AddIndexData is successful. If that does not happen either, it
remains uninitialized and we jump to the exit label, where bin2hex will
attempt to handle that dangling pointer.