backup: add option to skip broken attachments
Add option to skip broken attachments (without skipping the item).

Because we do want to log an error in such cases _and_ account for errors
in the ugly backup statistics dict _and_ ideally do this without adding multiple
arguments to Item.dump (and do that again for other methods..), it seemed time
to try out at a more general approach:

- pyko can figure out the log itself (service.log or server.log), so do it.
- Add a general Service.stats dict to maintain error or other counters.

The change is not yet completely factored out, but we can generalize
further when a similar issue occurs.

References: KC-45