libserver: log a MySQL error string instead of internal const
When a MySQL error occurs, the internal constant number is logged
instead of the MySQL error string. This makes debugging a lot harder,
since a user has to lookup the KCERROR_DATABASE_ERROR definition.
Therefore, log the MySQL error string when a select/update or
connection to the database fails.