python-mapi: register MAPIUninitialize to be called at exit
Even though leaking once at initialization time is not so bad, this
avoids many valgrind errors.

We call gc.collect() first, because we probably want uninitialization
to be really the last step before exiting. if we do not do this, we
will get segfaults after running certain tests.

References: KC-95