installer: split static parts of ldap.cfg
In the config file parser, only one set of defaults can be specified.
It is not possible to have some options default to different values
based on another option. As such, there is no option in the LDAP plugin
to select just the backend (AD/LDAP), and have defaults for e.g.
"ldap_last_modification_attribute" based on the backend selection. This
would require a two-step config file loading, which the parser does not
have.

As such, backend-specific options were traditionally put into the
config file itself. That however runs afoul of changed defaults, since
a config file, such as ldap.cfg, is considered to carry important
site-local overrides only. "ldap_last_modification_attribute" is not
something that is very site-specific, but rather backend-specific.

The backend values are therefore put into separate files, placed in
/usr (read-only by definition), to be sourced from a ldap.cfg in
/etc/kopano, and ldap.cfg can then override specific options if need
be.

References: KC-229