installer: avoid installation of ldap.cfg
[ldap.cfg was not installed beforehand (8.1) anyway and, for the
purposes of `make install`, was a new file just introduced 2 commits
ago and now we are pulling it again.]

Config files should not be tracked by the package manager. To this end,
our rpm/deb shall not install /etc/kopano/ldap.cfg. If not installed by
an rpm/deb package, there is no need to install it for source builds,
either. Therefore, remove it from $config_files. It will still be
installed to the exapmle-config directory and that is the only location
it should be.

References: KC-229