pyko: Add store.favorites()
Favorites are a MAPI Message which is stored in PR_COMMON_VIEWS
associated messages folder. The special properties set on this message
contain the store and folder entryid which can be used to open the
'favorite' folder.

This commit does not handle favorites which are not in your own store,
but for example are from a shared store.