backup: fix daemonization as user
Previously, we merged the foreground and daemonized cases into a single
"DaemonContext", but this does not work when running as a non-root
user. It is not setuid(my_own_uid) which fails, but setgroups([]),
which DaemonContext unconditionally calls and which is a privileged
operation even if the intent is to reduce the privilege level.

It only affects backup (and similar command-line utilities, so not
really services), an option to "not run as a service" is added.

References: KC-215