server: make allocator switchable and ensure it is loaded first
Because the dynamic runtime linker processes files breadth-first, the
malloc calls in the kopano-server process image get resolved to
libc.so's malloc (since that is directly referenced from kopano-server
via a DT_NEEDED entry) rather than libtcmalloc which is only referenced
indirectly via libkcserver.so.

Adding -ltcmalloc_minimal to all executables is not a fix deemed
suitable, because the linker might be operating in "--as-needed" mode
and might just effectively ignore the -l if the program is not using
malloc.

Which brings us to forced library loading through LD_PRELOAD. Benefits
include that different allocator libraries can now be chosen at
runtime, such as glibc, jemalloc or the full tcmalloc.

References: KC-196, KC-216, KC-219