admin: do not pass NULL to std::cout
kopano-admin can be triggered to pass NULL into std::cout, which is not
a good thing to do.

	kopano-admin -u $(echo -en '\x04') --purge-deferred

The validation for -u does not succeed, but the inputValidator object
is not reset, causing .Failed() to return true for all subsequent
arguments and always invoking cout << optarg. Choosing a no-argument
option causes optarg to be NULL.

CID 123155 (#1 of 1): Dereference after null check (FORWARD_NULL)
196. var_deref_model: Passing null pointer optarg to operator <<,
which dereferences it. [Note: The source code implementation of the
function has been overridden by a builtin model.]

References: CID-123155, KC-91