libserver: load libs3 via dlopen
By removing -ls3 from the executable, the problematic non-MT symbols
which may trickle in from libldap via e.g. libs3.so.2 → libcurl.so.4 →
libldap-2.4.so.2 will stay out of the process image if S3 is not
actually used.

The use of RTLD_LOCAL also seems helpful in that the ldap symbols won't
be added to the global resolution namespace, which means you can now
mix, if you are brave enough, S3 and LDAP.

References: KC-64