pyko: Add restriction class
This introduces a class which makes it easier to write MAPI Restrictions
they can be written as normal strings such as "subject == 'Python'".
PLY is used to parse the restriction string and to convert it to a MAPI
restriction.

Folder.items now takes a restriction as argument, so that items can be
easily filtered.

Item.match introduces the possibility to check if an
item matches a restriction by adding an and restriction on the
PR_SOURCE_KEY of the message.

The Rule class has a new restriction property which returns the
restriction of the rule.