php7: cure crash due to short-lived zvals
In PHP7, reference counting has been pushed from the zval struct down
to the object types (zend_object, zend_string), and it appears that
zvals no longer stay around after their scope ends. The zval pointed to
by "objImportContentsChanges" goes away, so the pointer is invalid.
This caused a crash. The object pointed to by
objImportContentsChanges->value.obj stays around however, so make use
of that. Valgrind (launched on php-fpm) certainly seems to be happy
with that.

References: KC-259