inetmapi: no more exceptions for normal code flow
All occurrences of try..catch blocks with respect to header->findField
have been removed. This is due to changes in vmime [commit
96077ce7e6125cdba959925d0a2b7b1f8ee54046] where findField will return
NULL in case the field is missing, instead of throwing an exception. We
will now check if the field exists first and then do whatever is needed
in case the field is present or missing.

References: KC-263
Signed-off-by: Bo Simonsen <bo@geekworld.dk>