build: do not depend on libtcmalloc_minimal.pc
While PKG_CHECK_MODULES can, in the absence of libtcmalloc_minimal.pc,
be overridden to succeed by manually setting TCMALLOC_CFLAGS,
PKG_CHECK_EXISTS always tries to locate the file and cannot be swayed
by command-line arguments.

This resolves a testsuite failure since there is a (new) check for
statistics containing a "tc_allocated" key (but which is only returned
if the server was compiled to return this info!).