mapi: switch to virtual inheritance and proper casting
When IID_IUnknown is requested, we must use an upcast to IUnknown, not
an upcast to some intermediate base and hope it has the same address.
Because same bases appear in the multiple inheritance chain, make use of
virtual inheritance so there is just one base.
This fixes UBSAN warnings like

../m4lcommon/include/kopano/mapi_ptr/mapi_object_ptr.h:207:4:
runtime error: member call on address 0x608000003ea0 which does not
point to an object of type 'IUnknown'
0x608000003ea0: note: object is of type 'M4LProfSect'
 0200804c  c0d61f26057f0000  70d71f26057f0000  02000000bebebebe
           ^~~~~~~~~~~~~~~~
           vptr for 'M4LProfSect'

References: KC-269