treewide: offer robust version of REGISTER_INTERFACE macro
REGISTER_INTERFACE did not perform any check whatsoever that the
2nd argument was in fact a class pointer that matches the GUID.

REGISTER_INTERFACE2 is hereby added, and which does this compile-time
check. The 3-argument version, REGISTER_INTERFACE3, is used when the
class name is different from the GUID name, but otherwise fitting.

Most uses of REGISTER_INTERFACE were converted over. Those that were
not are an indication of where the code breaks the API promise. The
ZCMAPIProp class for example tries to support IID_IMailUser, but since
there is no IMailUser in its class hierarchy, its return value is
questionable.

References: KC-235