spooler: avoid seeing yourself as a sender of a NDR
Non-delivery reports are generated by kopano-spooler, not the user,
therefore remove the user as a putative sender.

"SYSTEM" is a bad pick for sender, too, because:

1. SYSTEM is hidden by default, meaning that lookups by name return
no result (e.g. in WebApp) when creating new blank emails and the
name "SYSTEM" is underlined in red.

2. Sending *replies* curiously does resolve the SYSTEM user (likely
because it takes a shortcut of reusing PR_SENDER_ENTRYID). This
inconsistent user experience means it is better not to use that
account.

3. No admin who leaves the SYSTEM account untouched and at its
defaults would ever look at its inbox.

References: KC-309