common: abolish use-after-delete
Because “C++ explicitly allows an implementation of delete to zero out
an lvalue operand”[1], "this" should not be passed around anymore, even
if popular implementations retain the pointer value. As there is an
ordering imposed by the virtual destructor (e.g. ~ECMAPIFolder still
tries to use the list), the unlink has to remain at its place after
delete, which mandates a copy of the "this" pointer value is made.

[1] http://www.stroustrup.com/bs_faq2.html#delete-zero

References: KC-23, KC-235