libserver: make S3 implementation work with AWS4-HMAC-SHA256
We use the code from https://github.com/bji/libs3/pull/50 to make
AWS4-HMAC-SHA256 work. Furthermore, there was a problem in the
implementation when creating objects in S3: The size was set to the
contentLength response, but it may be missing (on PUT requests).
Therefore, we now do a check if it is present. The region configuration
parameter was also added.

References: KC-170