build: dissolve libkcsoap{client,server}
These two libraries have a limited number of users; libkcsoapserver for
example is only used by kopano-server and might as well just be
built-in directly so that we do not have to bother with symbol exports
at all. Rearrange the file lists in the Makefile. (Externally
determined with the help of graphviz's dot.)