inetmapi: recognize iCal descriptions as a body (1/2)
When a iCalendar object with a description was followed by a text MIME
part, that text part would, under certain conditions (no text part
prior to the iCal object), override the body already set by the iCal
description.

This happened because the iCal dissection function failed to advance
the internal state to indicate that body text was read.

References: KC-138