inetmapi: rework ASCII upgrades
"default_charset" is a misleading name choice. There are limited
instances where mail may be interpreted according to admin rules (say,
with cset EBCDICUS). In many cases, RFCs prescribe or strongly
recommend some default that is to be used, and this is how inetmapi/
was transmogrified from ZCP 7.2.3 onwards.

The real intent of the "default_charset" option however was a
admin-configurable reintepretation of ASCII parts under an
ASCII-derived encoding. The in-source code option is renamed to reflect
that purpose and this charset upgrading is (re)implemented.

The config option retains its name, for now at least.

References: KC-294