hl: make memory_ptr reset explicit
mapi_memory_ptr::operator& (and memory_ptr::operator&) would implicitly
reset _m_ptr. That is principally intentional, since it is expected
that a function wanting a pointer to a storage location would want to
overwrite the value in that location anyway; so, to not leak memory,
the _m_ptr had to be reset.

In the immediate future, we would like to use memory_ptr in more
places. With some functions like PublishFreeBusy::HrMergeBlocks
however, the implicit reset is undesired as much as a x=NULL statement
would be before realloc(x, …).

As such, turn the reset into an explicit operation. Let's bind this to
memory_ptr::operator~ for now. All current callers of operator& expect
the reset, so all &x calls are to be turned into &~x. To find all &x,
memory_ptr::operator& is marked private and returning void, which will
flag and identify all the callsites with compile errors (which we can
then edit, as is done). To allow & only after ~, a special proxy class
is used. (This might go away soon again.)

References: KC-329