common: handling segfault on non-writable logfile
In case the logger is constructed with a filename that cannot be
written to (i.e. permissions), it would segfault. The setup of file
handling functions has been split into different functions (i.e. log to
stderr, log to file, log to compressed file).

Also we use std::string for logname class member.