libicalmapi: fix searching for timezone
Previously, we picked the first STD/DST timezone definition found,
which was wrong because timezone definitions can change over time, and
you do not want to be using a definition from 1945, for example.

So we changed this to pick the first timezone definition matching the
current date. That is also not optimal, because this excludes
near-future timezone definitions, such as the DST starting in six
months from now.

We now improve things further by including the first timezone
definition going into the future. If there is no timezone definition
for the current date, pick the next best thing™. If there is only one
definition, it is back to the old situation.

References: KC-313