mapi: set MAPI_DIM to make a flexible array
aulPropTag and others is meant to be used as a flexible array.
Because old compilers did not support that, people used T[1]
as its type. With an explicit size given however, checkers
will flag apparent runoffs.

ICalToMAPI.cpp: warning: array index 43 is past the end of the array
(which contains 1 element) [-Warray-bounds]
	lpsPropValue[0].ulPropTag = CHANGE_PROP_TYPE(m_lpNamedProps->
	aulPropTag[PROP_TOATTENDEESSTRING], PT_UNICODE);
	           ~~~~~~~~~~~~~~~~~~~~~~

Contemporary compilers support flexible arrays, so make use of it to
quench of the warning, and maybe other assumptions/optimizations the
compiler was able to make for T[1].

References: KC-236