backup, search: replace python interpreter in #! line
backup and search contains a boilerplate python script for starting the
script (invoking main in __init__). These are dependent on the
interpreter, and since we gain initial support for python3 if
kopanocore is compiled with python3 these files would need to reference
python3. This commit uses autoconf to generate the scripts from the .in
file having @PYTHON@ from autoconf substituted.