server: abolish signal handling thread
The use of a signal thread has had the following drawbacks:

* Ctrl-C/SIGINT in gdb would not work: Because the signal is blocked,
it never reaches the process (and as such, a ptracer won't be notified
either), so gdb could not be halted. (KC-171)

Blocking signals is already a sufficient measure to direct signals to a
specific thread on Linux.

* Locking would be needed to manipulate the listening sockets' SSL
context. (KC-301)

References: KC-171, KC-301