inetmapi: set initial PR_MESSAGE_SIZE
At the current state, rule processing of messages based on size does
not work. Even the interval [0;<large number>] does not work due to
the absence of the PR_MESSAGE_SIZE attribute. In the current
implementation, the PR_MESSAGE_SIZE is computed on save (see
StorageUtil.cpp). This change will set that attribute initially to
the size of the RFC2822 message. That is ok according to MAPI
documentation; PR_MESSAGE_SIZE is an estimate of the message size. We
will allow setting PR_MESSAGE_SIZE until the message is saved, after
which we will not accept modifications.

References: KC-150