spooler: remove "for" in Received header
When a message has multiple recipients, the same recipient was added
to the "for" part of the received header for all recipients. It is
allowed by the RFC2822 to not include the recipient in the Received
header. We put the code for adding the Return-Path and Received
header into AddMiscHeaders to simplify the HandlerLMTP function.