treewide: make use of std::unique_ptr (1/2)
By using a managed pointer like unique_ptr, the explicit "delete" at
the end of the function is unnecessary because that will be taken care
of by ~unique_ptr already. That in turn shrinks the number of lines
after the exit: label, often to just a return, which then allows
replacing all gotos with just return.

References: KC-329