client: implement Kerberos/GSSAPI support
This implements GSSAPI authentication in libkcclient which allows it
to be used with Kerberos authentication in particular.

SSO authentication is always tried (just as before), but now has a
chance to succeed if a credential cache with a valid TGT or kopano
service ticket can be found. This allows all standard clients (e.g.
kopano-set-oof) to be used with GSSAPI (Kerberos) authentication.

This also allows use of Constrained Delegation since no special
treatment of that case is required — GSSAPI handles all the details
internally, provided it finds the necessary prerequisites in the
default credential cache.

References: KC-396