gateway: replace old SRestriction code with ECRestriction
Coverity bemoans:

CID-158424 (#1-6 of 6): Resource leak (RESOURCE_LEAK)
75. leaked_storage: Variable lpExtraRestriction going out of scope
leaks the storage it points to.

The encompassing function is so huge that, before even bothering to
look for the exact lifetimes of lpRootRestriction and
lpExtraRestriction, rip it all out and replace it by the modern
ECRestriction-based restriction construction.

The old code also constructed structures in an inefficient way, by
doing AND(cond1, AND(cond2, AND(cond3, …))) rather than just AND(cond1,
cond2, cond3, …), which has now been fixed.