m4lcommon: re-evaluate copy modes for EC{Property,Content,Raw}Restriction
When ECRestriction member functions are invoked on a just-constructed
object, like this:

	ECPropertyRestriction(x, y, &pv).RestrictTable(...)
or e.g.
	ECContentRestriction(x, y, &pv).CreateMAPIRestriction(...)

then pv is guaranteed to be vaild and unmodified for the
RestrictTable() call to complete, so all those instances are converted
to use the "cheap" copy mode.

In case a SPropValue struct is reused during build-up of a larger
restriction, like so:

	pv.ulPropTag = PR_BODY;
	pv.Value.lpszA = guid1.c_str();
	r += ECPropertyRestriction(x, y, &pv);
	pv.ulPropTag = PR_HTML;
	pv.Value.lpszA = guid2.c_str();
	r += ECPropertyRestriction(x, y, &pv);
	r.RestrictTable()

then "copy" is not enough and the "shallow" mode is to be used,
provided the constituents remain unchanged.

When the restriction is returned from a function, a full copy is
usually needed when objects go out of scope.