m4lcommon: constify input SPropValues for ECRestriction
EC{Raw,Content,Property}Restriction either makes a shallow/full copy,
in which the original SPropValue remains unchanged, or it makes a cheap
copy, in which it is supposed to be left unchanged too. Therefore, the
arguments should carry a "const" qualifier. Unfortunately, we have to
use a const_cast<> to funnel it into shared_ptr, but there is our
little promise to not free it.

References: KC-358