m4lcommon: facilitate top-down construction of ECRestrictions
To create an ECNotRestriction, one already needs to know the NOT's
subrestriction, in other words, a restriction tree needs to be built
from the bottom up. IMAP.cpp does a top-down build however and would
need overly much rewrite to move it to ECRestriction. So, do away with
ECRestriction's limitation and make top-down construction possible.