inetmapi: do not wrap (Content|Message)-Id
Outlook and maybe other MUAs has problems on Content-Id is being
wrapped (i.e. Content-Id: \n<...>).

In VMIME Content-Id and Message-Id is being handled identically, so we
added a feature for VMIME to force not wrapping these headers, and
with this commit we apply this feature during initialization of
inetmapi.

References: KC-361