common: trim all trailing whitespace on HTML-to-text conversion
Each </p> tag is transformed into \n\n which, at the end of a document,
leads to extraneous white lines, which is hereby stripped. This also
has the effect of removing white lines that would be (legitimately)
emitted if the input stream contained instances of <p>&nbsp;</p> at the
end. This change is arguably justified, because white lines at the end
of a mail only cause the user to needlessly spend time scrolling
downwards in the hope to find some more text (but which is not going to
be there).

If we did not strip it in the HTML-to-text converter, then (a) the
testsuite would have to be sprinkled with .rstrip() calls all over the
place to cater for the case, (b) repeated text->html->text->html->…
conversion would not be idempotent and lead to an ever-growing body.

References: KC-338