pyko: replace usage of MAPI.Util.AddressBook with service admin
pyko is perhaps the only user of MAPI.Util.AddressBook at this point,
and apparently it doesn't work as expected at some customer, whereas
kopano-admin works fine..

we rewrite the respective code to use the service-admin interface,
which kopano-admin also uses. so we are more consistent, the problem
at the customer should be resolved, and we may start to think about
removing MAPI.Util.AddressBook.

References: KC-492