gateway: enforce user and password checking when using the UNIX socket
When using the UNIX socket to connect to kopano-server and the
process is running as the same unix user, special privileges
are obtained by the client, so user and password are not checked.
This is useful for migration purposes, however an admin may be
able to misconfigure the system leaving it open for everybody
who knows the username of a user. We want to enforce checking
per default unless it is explicitly turned off (by bypass_auth).

References: KC-490