gateway: allow caching of folder list
Added option "imap_cache_folders_time_limit". When it is greater
than zero, the list of folders will be cached for N seconds given
the option is N. This is to ensure faster refresh times when
checking the entire hierarchy for new mails. When a user
issues a LIST command the folder list is being refreshed,
as well for DELETE, RENAME and CREATE.

References: KC-490