gateway: optimize UID FETCH X:Y (UID FLAGS)
It is common to see the following query by MUAs (e.g. Evolution)
to check if the folder contains the same messages as previously
seen. The query did result in several MAPI calls eventhough we
had all the required data in memory which is being utilized
in this commit.

References: KC-490