php7-ext: defining zend_isnan to avoid compilation failure on gcc version < 6.0
std::isnan was introduced for C++11 however the
php C headers has not been updated to support the
existence of that function. When using gcc6
this does not result in an error, however for
older versions it does. This commit ensures that
we can compile php7 support for older versions
than gcc version 6.