common: plug memory leak near memory_ptr<SRowSet> instances
There are two uncanny memory_ptr<SRowSet> variables in the code. When
they go out of scope, only MAPIFreeBuffer is invoked, but most SRowSets
warrant freeing with FreeProws — these two included — because
lpRowSet->aRow[n].lpProps is not linked to the SRowSet by way of
MAPIAllocateMore.

IOW, there is a memory leak. To resolve this, memory_ptr is switched to
use a deleter functor that acts specially on SRowSet.

References: KC-441