pyko: limit open message store cache
As observed for a very large multiserver setup (10000+ users), there
seems to be a limit to the number of MAPI stores we can keep open.
For this particular setup, that is reproducibly very close to 1000
stores.

Since there's no real need to cache 1000's of stores, we just use
the nice new Python 3 functools.lru_cache decorator to limit the pyko
store cache to a max of 128 stores. That means we need to add a
2.x backport, so we do.

As tested on the setup, we can now loop over 10284 users without a
problem.

References: KC-503