libserver: fixed refactoring mistake on getting server details
In commit bp/8.3.0-f735b5f6, the check of EC_SERVERDETAIL_NO_NAME
flag was changed. The condition was inverted, so in case the flag
was set the name would be copied, but it should be the other way
around. This commit corrects that, and fixes crash in dagent (and
probably other components) in a multi-server setup.

References: KC-531