mapi: silence Coverity TOCTOU complaint
CID-159538 (#1 of 1): Time of check time of use (TOCTOU)
6. fs_check_call: Calling function stat to perform check
on path->c_str()
7. toctou: Calling function opendir that uses path->c_str()
after a check function. This can cause a time-of-check,
time-of-use race condition.

It does not matter. If someone changed the filesystem object between
stat and opendir, opendir is a sufficient call to ensure the thing
really is a directory. Drop the useless stat call.

References: KC-236