libserver: turn DB_RESULT into a wrapping class
There is already a DB_RESULT_AUTOFREE class in plugins/DBBase.h, but
it suffers from problems:

1. DB_RESULT_AUTOFREE r = y; and r = y; set the database,
   not the result, which makes statements like "r = y; if (r == y)"
   not behave in their usual fashion.

2. The class does not guard against someone calling d->FreeResult(r).
   The subsequent ~DB_RESULT_AUTOFREE invocation could then free a
   dangling pointer if FreeResult was not followed by an explicit
   r = NULL.

3. The database pointer must be manually set beforehand
   (and hope it is right).

Thus, the route of transforming DB_RESULT_AUTOFREE will not be pursued.
Instead, we will have a pitch at DB_RESULT.

In this first step, DB_RESULT is turned from a typedef into a class
that behaves the same.

References: KC-532