pyko: set $HOME to a new temp dir after daemonizing
When daemonizing, one might expect the $HOME variable to change
according to the user that is switched to. So that for example,
a configuration file can be created here (by some conversion tool
called by search, such as w3m). Unfortunately, in our case, $HOME
is not writable, so we create a tempdir instead.

References: KC-331