swig: use memory_ptr for mapilists
Further use memory_ptr to avoid MAPIFreeBuffer calls for mapilists.
Avoid interfering with special-casing for adrlists, or we may
double-free because of calling FreeProws.

This special-casing is a bit nasty, and could use a bit of a cleanup
in the future. For example by replacing it with non-overlapping
patterns.

References: KC-541