pyko: fall back to search-key to determine recipient email address
For a recipient with addrtype ZARAFA, when the respective user is
deleted or just non-existent (eg after a migration), we cannot resolve
the user anymore, and determine their email address. Fortunately,
the email address is usually still stored as part of PR_SEARCH_KEY.
So we fall back to that for now. Long-term we probably want to always
store recipient SMTP addresses (webapp/z-push/backup/migrations..).

References: KC-566