pyko: make prop() methods raise NotFoundError
As promised, we should raise a more readable exception when a
property cannot be found, and be consistent with other such
situations, such as when a user or item cannot be found.

So this is "the" incompatible change for 8.4 as far as I'm
concerned, and people may need to adjust their scripts for this.