server: demote readlink log message to debug
In a chroot where /proc is not mounted and no mysql host is
reachable, the console output is a bit curious with a default install
(log_level=3):

kc_reexec_with_allocate: readlink: No such file or directory
Terminated

This is because ECLogger is not yet initialized to log to a file, but
then at some point is. (It later dies because of the unreachable
mysql.) Demote the log message to debug to make it go away in normal
operation, because it does not matter anyway.