libserver: replace goto exit by return
"goto exit" clauses can be transformed to a direct return when the
code at the exit: label will effectively be skipped / when it is a
no-op for the particular case that led up to the goto statement.

Coverity is expected to report new warnings about tautological
subclauses (in if constructs) — this will be dealt with when those
reports arrive.