libserver,etc: devirtualize KDatabase::GetNumRows calls
From an abstract OO perspective, db->GetNumRows(result) is a strange
way to query a property of the result. Asking the object that has the
property in question associated directly, e.g.
result->get_num_rows(), is a lot more natural.

Given DB_RESULT is its own class (rather than just a typedef) since
commit bp/8.3.0-754-g087d24ad , this OO API oddity can now be
rectified. DB_RESULT even has a reference/pointer to the database it
is associated with (just in case this is needed for some mysql
calls), so a result object has everything to answer any requests for
result properties on its own, without requiring KDatabase::
functions.

In addition, there should also be a miniscule performance improvement
to be expected by this change, as KDatabase is a class with virtual
methods and DB_RESULT is not.