libserver: respond with no data for getLicenseAuth request
I do not remember why this function was made to return what is
essentially 24 NUL bytes to the network. The ZCP client cannot do
anything useful with that and so refuses to open archived/stubbed
messages.

Just return a zero-length block instead, like it is already done in
ns__logon(), where the lpsResponse->sLicenseResponse field is not
touched at all and remains in its default empty state (size=0).

References: KA-2