swig: open items lazily on Folder.items() call
Some users may issue the call: list(store.inbox.items())[-1]

Before this commit, this call would open the table, traverse it
and then call OpenEntry on all items. Having a folder with 1000+
items this may be very slow just to get one element. Now we
traverse the table and just create the item with the entryid.
Whenever a method/property is invoked then the item is opened.