inetmapi: do not force HTML when use_tnef is set to "minimal"
The logic for detecting the body is as follows:

- if the RTF is converted from HTML, use the HTML RTF stream
- if the RTF is text, keep that
- otherwise use the real RTF

This works fine, except for when the TNEF mode is set to minimal.
Here, we want to use the HTML stream or just plaintext. An attempt to
introduce this behavior unfortunately sent HTML everytime, even for
plain text emails. This commit fixes that.

References: KC-664