mapitime: add additional timing tests
The "lsr" test function only measures the actual (de)serialization
speed, so it is purely CPU-bound and not very comparable to the
server-involving test functions like "lilo".

So we need a lsr that does issue a network call, and this is what the
"lsr+ping" variant is about. For comparison to plain login,
a "ping" function is added too.

ping: 128 op/s, lsr+ping: 180 op/s.

References: KC-67