treewide: audit OpenEntry calls for right pointer types
The *::OpenEntry calls, when used with lpInterface=nullptr, can
return an arbitrary pointer type. The ulObjectType argument indicates
which one you got. However, this is not helping at all for object
types the code does not handle or does not know about. Casting that
random-typed pointer to IUnknown and using that result is undefined
behavior.

	void *unk;
	x->OpenEntry(…, nullptr, &objtype, &unk);
	if (objtype == MAPI_ABCONT) {
		auto ab = reinterpret_cast<IABContainer *>(unk);
		ab->foobar();
		ab->Release();
	} else {
		/* BOOM, wrong — the type might have been a IDistList*. */
		reinterpret_cast<IUnknown *>(unk)->Release();
		return MAPI_E_UNKNOWN_ENTRYID;
	}

The UBSAN reports look something like:

mapix.cpp:1158:17: runtime error: member call on address 0x000002203588
which does not point to an object of type 'IMSProvider'
0x000002203588: note: object is of type 'ECMSProviderSwitch::xMSProvider'
 00 00 00 00  78 bc 09 d8 8f 7f 00 00  04 00 00 00 00 00 00 00
              ^~~~~~~~~~~~~~~~~~~~~~~
              vptr for 'ECMSProviderSwitch::xMSProvider'

Now, if one cannot cast to the right type, it is impossible to call
any of the non-static methods, which means the object cannot be
disposed of, either.

The only solution I see here is to explicitly request a particular
interface at first and ask again for another interface afterwards.

	IUnknown *unk;
	x->OpenEntry(…, IID_IUnknown, &objtype, &unk);
	if (objtype == MAPI_ABCONT) {
		unk->QueryInterface(IID_IABContainer, &ab);
		ab->foobar();
	} else {
		ret = MAPI_E_UNKNOWN_ENTRYID;
	}
	unk->Release();

References: KC-673